+++ src/api.ts
@@ -1,0 +1,1 @@
+export { usePushError } from "./contexts/error";