+++ src/contexts/ui.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 
 type UiState = {
+  errors: boolean;
   pending: number;
   settings: boolean;
 };
@@ -8,6 +9,7 @@ type UiState = {
 type UiContext = UiState & {
   pushLoader: () => void;
   popLoader: () => void;
+  toggleErrors: () => void;
   toggleSettings: () => void;
 };
 
@@ -15,6 +17,7 @@ export const UiContext = React.createContext({} as unknown as UiContext);
 
 const UiProvider: React.FC = ({ children }) => {
   const [state, setState] = React.useState<UiState>({
+    errors: false,
     pending: 0,
     settings: false,
   });
@@ -25,6 +28,8 @@ const UiProvider: React.FC = ({ children }) => {
         setState((state) => ({ ...state, pending: state.pending + 1 })),
       popLoader: () =>
         setState((state) => ({ ...state, pending: state.pending - 1 })),
+      toggleErrors: () =>
+        setState((state) => ({ ...state, errors: !state.errors })),
       toggleSettings: () =>
         setState((state) => ({ ...state, settings: !state.settings })),
     }),