+++ src/views/Root.tsx
@@ -1,4 +1,5 @@
 import React from "react";
+import ErrorProvider from "../contexts/error";
 import TimeProvider from "../contexts/time";
 import UiProvider from "../contexts/ui";
 import IntlProvider from "../locales/IntlProvider";
@@ -6,13 +7,15 @@ import App from "./App";
 
 const Root: React.FC = () => (
   <React.StrictMode>
-    <UiProvider>
-      <IntlProvider>
-        <TimeProvider>
-          <App />
-        </TimeProvider>
-      </IntlProvider>
-    </UiProvider>
+    <ErrorProvider>
+      <UiProvider>
+        <IntlProvider>
+          <TimeProvider>
+            <App />
+          </TimeProvider>
+        </IntlProvider>
+      </UiProvider>
+    </ErrorProvider>
   </React.StrictMode>
 );