+++ src/views/shared/modal/Modal.tsx
@@ -1,10 +1,16 @@
 import React from "react";
 import "./Modal.css";
 
-const Modal: React.FC = ({ children }) => {
+type Props = {
+  onClose: () => void;
+};
+
+const Modal: React.FC<Props> = ({ children, onClose }) => {
   return (
-    <div className="Modal-container">
-      <div className="Modal">{children}</div>
+    <div className="Modal-container" onClick={onClose}>
+      <div className="Modal" onClick={(event) => event.stopPropagation()}>
+        {children}
+      </div>
     </div>
   );
 };